增加了模板组件

+++ pre_term/App/views/file.py
@@ -15,24 +15,6 @@ from App.models import *
 from App.views_helper import send_email_activate
 
 
-def create_file(request):
-    if request.method == 'GET':
-        return render(request, 'file/create_file.html')
-    elif request.method == 'POST':
-        content = request.POST.get('content')
-        title = request.POST.get('title')
-        file = File()
-        file.title = title
-        file.content = content
-        file.creator = request.user.u_username
-        file.save()
-        personal_record = Personal_record()
-        personal_record.user = request.user
-        personal_record.files = file
-        personal_record.save()
-        return redirect(reverse('app:file_info', args={file.id}))
-
-
 # 修改文件
 def change_file(request, file_id):
     if request.method == 'GET':
@@ -68,7 +50,6 @@ def my_files_list(request):
         files = File.objects.filter(personal_collection__user=user).filter(is_delete=False).order_by('-id')
     else:
         files = []
-    # files = File.objects.all().order_by('-id')
 
     page = int(request.GET.get("page", 1))
     perpage = int(request.GET.get('perpage', 10))
@@ -120,16 +101,43 @@ def file_info(request, file_id):
 
 
 # 文件删除
-def delete_file(request, file_id):
-    file = File.objects.get(pk=file_id)
-    file.is_delete = True
-    file.save()
-    deletedate = delete_date()
-    deletedate.file = file
-    deletedate.user = request.user
-    deletedate.save()
-
-    return redirect(reverse('app:my_files_list'))
+def delete_file(request):
+    id = request.POST['id']
+    file = File.objects.get(pk=id)
+    if file.is_delete:
+        return JsonResponse(data={"msg": "文件已删除", 'status': 1})
+    try:
+        is_team = Team_record.objects.filter(files=file)
+        if is_team.exists():
+            team = is_team.first().team
+            level = Team_relation.objects.filter(team=team).filter(user=request.user).first().level
+            if level >= 2:
+                file.is_delete = True
+                file.save()
+                date = delete_date()
+                date.file = file
+                date.user = request.user
+                date.save()
+                return JsonResponse(data={"msg": "删除团队文档成功", 'status': 0})
+            else:
+                return JsonResponse(data={"msg": "删除团队文档权限不足", 'status': 1})
+        else:
+            is_person = Personal_record.objects.filter(files=file).filter(user=request.user)
+            if is_person.exists():
+                if is_person.first().is_creator:
+                    file.is_delete = True
+                    file.save()
+                    date = delete_date()
+                    date.file = file
+                    date.user = request.user
+                    date.save()
+                    return JsonResponse(data={"msg": "删除个人文档成功", 'status': 0})
+                else:
+                    return JsonResponse(data={"msg": "删除个人文档权限不足", 'status': 1})
+            else:
+                return JsonResponse(data={"msg": "删除个人文档失败", 'status': 1})
+    except:
+        return JsonResponse(data={"msg": "删除个人文档失败", 'status': 1})
 
 
 # 修改日志
@@ -191,3 +199,184 @@ def file_search(request):
         files = File.objects.filter(
             Q(title__icontains=content) | Q(content__icontains=content) | Q(creator__icontains=content))
         return render(request, 'file/file_search.html', context={'files': files})
+
+
+# 获取模板
+def get_templetes(request):
+    templetes = Template.objects.all()
+    print(templetes)
+    res = []
+    for templete in templetes:
+        dic = {
+            "id": templete.id,
+            "content": templete.content,
+            "title": templete.title,
+            "score": templete.score,
+            "accept_num": templete.accept_num,
+            "img": str(templete.img),
+        }
+        res.append(dic)
+    data = {
+        'msg': '获取全部模板成功',
+        'templetes_list': res
+    }
+    return JsonResponse(data=data)
+
+
+# 给模板评分
+def grade_templetes(request):
+    id = int(request.POST['id'])
+    score = float(request.POST['score'])
+
+    templete = Template.objects.get(pk=id)
+    if templete:
+        if score <= 5 and score >= 0:
+            if templete.accept_num == 0:
+                templete.score = score
+            else:
+                templete.score = (templete.score + score) / templete.accept_num
+            data = {
+                'msg': '评分成功',
+                'status': 0
+            }
+        else:
+            data = {
+                'msg': '请在0-5范围内评分',
+                'status': 1
+            }
+    else:
+        data = {
+            'msg': '该模板不存在',
+            'status': 1
+        }
+    return JsonResponse(data=data)
+
+
+# 新建文档
+def create_file(request):
+    try:
+        templete_id = int(request.POST['templete_id'])
+        title = request.POST['title']
+        content = request.POST['content']
+        team_id = int(request.POST['team_id'])
+        if templete_id == 0 and team_id == 0:
+            file = File()
+            file.creator = request.user.u_username
+            file.save()
+
+            personal_record = Personal_record()
+            personal_record.user = request.user
+            personal_record.files = file
+            personal_record.save()
+            data = {
+                'msg': '未使用模板的个人文档创建成功',
+                'status': 0,
+                'id': file.id
+            }
+        elif templete_id != 0 and team_id == 0:
+            templete = Template.objects.get(pk=templete_id)
+            file = File()
+            file.creator = request.user.u_username
+            file.title = templete.title
+            file.content = templete.content
+            file.save()
+            templete.accept_num = templete.accept_num + 1
+            templete.save()
+
+            personal_record = Personal_record()
+            personal_record.user = request.user
+            personal_record.files = file
+            personal_record.save()
+            data = {
+                'msg': '使用模板的个人文档创建成功',
+                'status': 0,
+                'id': file.id
+            }
+        elif templete_id == 0 and team_id != 0:
+            file = File()
+            file.creator = request.user.u_username
+            file.save()
+
+            team_record = Team_record()
+            team_record.team = Team.objects.get(pk=team_id)
+            team_record.files = file
+            team_record.save()
+            data = {
+                'msg': '未使用模板的团队文档创建成功',
+                'status': 0,
+                'id': file.id
+            }
+        elif templete_id != 0 and team_id != 0:
+            templete = Template.objects.get(pk=templete_id)
+            file = File()
+            file.creator = request.user.u_username
+            file.title = templete.title
+            file.content = templete.content
+            file.save()
+            templete.accept_num = templete.accept_num + 1
+            templete.save()
+
+            team_record = Team_record()
+            team_record.team = Team.objects.get(pk=team_id)
+            team_record.files = file
+            team_record.save()
+            data = {
+                'msg': '使用模板的团队文档创建成功',
+                'status': 0,
+                'id': file.id
+            }
+        else:
+            data = {
+                'msg': '新建失败',
+                'status': 1
+            }
+    except:
+        data = {
+            'msg': 'wrong'
+        }
+    return JsonResponse(data=data)
+
+
+# 个人文档给协作者授予权限
+def grant_power(request):
+    try:
+        u_id = int(request.GET['u_id'])
+        file_id = int(request.GET['id'])
+        comment = int(request.GET['comment'])
+        change = int(request.GET['change'])
+        personal_record = Personal_record.objects.filter(user_id=u_id).filter(files_id=file_id).first()
+
+        if comment == 1:
+            personal_record.comment = True
+            personal_record.save()
+        elif comment == 0:
+            personal_record.comment = False
+            personal_record.save()
+        else:
+            data = {
+                'msg': 'power wrong',
+                'status': 1
+            }
+            return JsonResponse(data=data)
+        if change == 1:
+            personal_record.change = True
+            personal_record.save()
+        elif change == 0:
+            personal_record.change = False
+            personal_record.save()
+        else:
+            data = {
+                'msg': 'power wrong',
+                'status': 1
+            }
+            return JsonResponse(data=data)
+        data = {
+            'msg': '修改权限成功',
+            'status': 0
+        }
+    except:
+        data = {
+            'msg': 'wrong',
+            'status': 1
+        }
+    return JsonResponse(data=data)
